+++ Demos/Benchmark/BenchmarkDemo.cpp
@@ -26,8 +26,11 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE. */
 #include <numeric>
 #include <algorithm>
 #include <iterator>
+#include <cmath>
+#include <fstream>
 
 #include "silkit/SilKit.hpp"
+#include "silkit/SilKitVersion.hpp"
 #include "silkit/services/all.hpp"
 #include "silkit/services/orchestration/all.hpp"
 
@@ -49,8 +52,7 @@ std::ostream& operator<<(std::ostream& out, std::chrono::nanoseconds timestamp)
 void PrintUsage(const std::string& executableName)
 {
     std::cout << "Usage:" << std::endl
-              << executableName 
-              << " [numberOfSimulationRuns]"
+              << executableName << " [numberOfSimulationRuns]"
               << " [simulationDuration]"
               << " [numberOfParticipants]"
               << " [messageCount]"
@@ -60,11 +62,17 @@ void PrintUsage(const std::string& executableName)
               << "\t--help\tshow this message." << std::endl
               << "\t--registry-uri\tThe URI of the registry to start. Default: silkit://localhost:8500" << std::endl
               << "\t--message-size\tSets the message size to BYTES. Default: 1000" << std::endl
-              << "\t--message-count\tSets the number of messages to be send per participant in each simulation step to NUM. Default: 50" << std::endl
+              << "\t--message-count\tSets the number of messages to be send per participant in each simulation step to "
+                 "NUM. Default: 50"
+              << std::endl
               << "\t--number-participants\tSets the number of simulation participants to NUM. Default: 2" << std::endl
-              << "\t--number-simulation-runs\tSets the number of simulation runs to perform to NUM. Default: 4" << std::endl
-              << "\t--simulation-duration\tSets the simulation duration (virtual time) to SECONDS. Default: 1s" << std::endl
-              << "\t--configuration\tPath and filename of the participant configuration YAML or JSON file" << std::endl;
+              << "\t--number-simulation-runs\tSets the number of simulation runs to perform to NUM. Default: 4"
+              << std::endl
+              << "\t--simulation-duration\tSets the simulation duration (virtual time) to SECONDS. Default: 1s"
+              << std::endl
+              << "\t--configuration\tPath and filename of the participant configuration YAML or JSON file. Default: empty"
+              << std::endl
+              << "\t--write-csv\tPath and filename of csv file with benchmark results. Default: empty" << std::endl;
 }
 
 struct BenchmarkConfig
@@ -76,6 +84,7 @@ struct BenchmarkConfig
     uint32_t messageSizeInBytes = 1000;
     std::string registryUri = "silkit://localhost:8500";
     std::string silKitConfigPath = "";
+    std::string writeCsv = "";
 };
 
 bool Parse(int argc, char** argv, BenchmarkConfig& config)
@@ -155,6 +164,7 @@ bool Parse(int argc, char** argv, BenchmarkConfig& config)
     parseOptional("--number-simulation-runs", config.numberOfSimulationRuns, asNum);
     parseOptional("--simulation-duration", config.simulationDuration, asNum);
     parseOptional("--configuration", config.silKitConfigPath, asStr);
+    parseOptional("--write-csv", config.writeCsv, asStr);
 
     //check unknown long options
     for (const auto& arg : args)
@@ -303,31 +313,56 @@ void ParticipantsThread(std::shared_ptr<SilKit::Config::IParticipantConfiguratio
 
 void PrintParameters(BenchmarkConfig benchmark)
 {
-
 #ifndef NDEBUG
     std::cout << "WARNING: The benchmark demo is executed in a DEBUG build configuration." << std::endl
               << "For more reliable timings, please use a RELEASE build configuration" << std::endl
-              << "of the SIL Kit library and the benchmark demo."
-              << std::endl;
+              << "of the SIL Kit library and the benchmark demo." << std::endl;
     std::this_thread::sleep_for(2s);
 #endif
 
     std::cout << std::endl
               << "This benchmark demo produces timings of a configurable simulation setup." << std::endl
-              << "<N> participants exchange <M> messages of <B> bytes per simulation step" << std::endl 
+              << "<N> participants exchange <M> messages of <B> bytes per simulation step" << std::endl
               << "with a fixed period of 1ms and run for <S> seconds (virtual time)." << std::endl
               << "This simulation run is repeated <K> times and averages over all runs are calculated." << std::endl
               << "The demo uses PubSub controllers with the same topic for the message exchange," << std::endl
               << "so each participant broadcasts the messages to all other participants." << std::endl
               << std::endl
-              << "Running simulations with the following parameters:" << std::endl << std::endl
-              << std::left << std::setw(38) << "- Number of simulation runs: " << benchmark.numberOfSimulationRuns << std::endl
-              << std::left << std::setw(38) << "- Simulation duration (virtual time): " << benchmark.simulationDuration << std::endl
-              << std::left << std::setw(38) << "- Number of participants: " << benchmark.numberOfParticipants << std::endl
-              << std::left << std::setw(38) << "- Messages per simulation step (1ms): " << benchmark.messageCount << std::endl
-              << std::left << std::setw(38) << "- Message size (bytes): " << benchmark.messageSizeInBytes << std::endl
-              << std::left << std::setw(38) << "- Registry URI: " << benchmark.registryUri << std::endl
+              << "Running simulations with the following parameters:" << std::endl
+              << std::endl
+              << std::left << std::setw(38) << "- Number of simulation runs: " << benchmark.numberOfSimulationRuns
+              << std::endl
+              << std::left << std::setw(38) << "- Simulation duration (virtual time): " << benchmark.simulationDuration
+              << std::endl
+              << std::left << std::setw(38) << "- Number of participants: " << benchmark.numberOfParticipants
+              << std::endl
+              << std::left << std::setw(38) << "- Messages per simulation step (1ms): " << benchmark.messageCount
+              << std::endl
+              << std::left << std::setw(38) << "- Message size (bytes): " << benchmark.messageSizeInBytes
+              << std::endl
+              << std::left << std::setw(38) << "- Registry URI: " << benchmark.registryUri
+              << std::endl
+              << std::left << std::setw(38) << "- Configuration: " << benchmark.silKitConfigPath
+              << std::endl
+              << std::left << std::setw(38) << "- CSV output: " << benchmark.writeCsv
               << std::endl;
+}
+
+template <typename T>
+std::pair<T, T> mean_and_error(const std::vector<T>& vec)
+{
+    const size_t sz = vec.size();
+    if (sz == 1)
+    {
+        return std::make_pair(vec[0], 0.0);
+    }
+
+    const T mean = std::accumulate(vec.begin(), vec.end(), 0.0) / sz;
+    auto variance_func = [&mean, &sz](T accumulator, const T& val) {
+        return accumulator + ((val - mean) * (val - mean) / (sz - 1));
+    };
+
+    return std::make_pair(mean, std::sqrt(std::accumulate(vec.begin(), vec.end(), 0.0, variance_func)));
 }
 
 /**************************************************************************************************
@@ -366,7 +401,7 @@ int main(int argc, char** argv)
         {
             std::vector<size_t> counters(benchmark.numberOfParticipants, 0);
             std::cout << "> Simulation " << simulationRun << ": ";
-            auto startTimestamp = std::chrono::system_clock::now();
+            auto startTimestamp = std::chrono::high_resolution_clock::now();
 
             std::vector<std::string> participantNames;
             std::vector<std::thread> threads;
@@ -377,8 +412,8 @@ int main(int argc, char** argv)
                 participantNames.push_back(participantName);
                 auto& counter = counters.at(idx);
                 idx++;
-                threads.emplace_back(&ParticipantsThread, config, benchmark, participantName,
-                                     participantIndex, std::ref(counter));
+                threads.emplace_back(&ParticipantsThread, config, benchmark, participantName, participantIndex,
+                                     std::ref(counter));
             }
 
             const auto systemControllerName = "SystemController";
@@ -399,7 +434,7 @@ int main(int argc, char** argv)
 
             systemControllerParticipant.reset();
 
-            auto endTimestamp = std::chrono::system_clock::now();
+            auto endTimestamp = std::chrono::high_resolution_clock::now();
             measuredRealDurations.emplace_back(endTimestamp - startTimestamp);
             auto totalCount = std::accumulate(counters.begin(), counters.end(), size_t{0});
             messageCounts.emplace_back(totalCount);
@@ -407,15 +442,32 @@ int main(int argc, char** argv)
             std::cout << " " << measuredRealDurations.back() << std::endl;
         }
 
-        const auto minMaxDuration = std::minmax_element(measuredRealDurations.begin(), measuredRealDurations.end());
-        const auto averageDuration = std::accumulate(measuredRealDurations.begin(), measuredRealDurations.end(), 0ns)
-                                     / measuredRealDurations.size();
+        std::vector<double> measuredRealDurationsSeconds(measuredRealDurations.size());
+        std::transform(measuredRealDurations.begin(), measuredRealDurations.end(), measuredRealDurationsSeconds.begin(),
+                       [](auto d) {
+                           return static_cast<double>(d.count() / 1e9);
+                       });
+
+        const auto averageDuration = mean_and_error(measuredRealDurationsSeconds);
+
         const auto averageNumberMessages =
             std::accumulate(messageCounts.begin(), messageCounts.end(), size_t{0}) / messageCounts.size();
 
-        // Byte to Mebi and ns to s
+        // Reoccuring factor from error propagation
+        const auto sigmaDurOverDurSqr = averageDuration.second / averageDuration.first / averageDuration.first;
+
+        // Byte throughput mean and error (Byte to Mebi and ns to s)
+        const auto throughputPrefac = averageNumberMessages * benchmark.messageSizeInBytes / 1024.0 / 1024.0;
         const auto averageThroughput =
-            averageNumberMessages * benchmark.messageSizeInBytes / 1024.0 / 1024.0 / averageDuration.count() * 1e9;
+            std::make_pair(throughputPrefac / averageDuration.first, throughputPrefac * sigmaDurOverDurSqr);
+
+        // Message rate mean and error
+        const auto averageMsgRate =
+            std::make_pair(averageNumberMessages / averageDuration.first, averageNumberMessages * sigmaDurOverDurSqr);
+
+        // Speedup mean and error
+        const auto averageSpeedup = std::make_pair(benchmark.simulationDuration.count() / averageDuration.first,
+                                                   benchmark.simulationDuration.count() * sigmaDurOverDurSqr);
 
         if (benchmark.numberOfSimulationRuns > 1)
         {
@@ -426,18 +478,81 @@ int main(int argc, char** argv)
             std::cout << std::endl << "Result of the simulation run:" << std::endl << std::endl;
         }
 
-        std::cout << std::left << std::setw(38) << "- Realtime duration: " << averageDuration << std::endl
-                  << std::left << std::setw(38) << "- Total number of messages: " << averageNumberMessages << std::endl
-                  << std::left << std::setw(38) << "- Message rate per participant: " << benchmark.messageCount * 1000 << " #/s (virtual time)" << std::endl
-                  << std::left << std::setw(38) << "- Throughput: " << averageThroughput << " MiB/s" << std::endl;
+        // Stream helper to combine value and unit to use it with std::setw as a whole
+        std::ostringstream averageDurationWithUnit;
+        averageDurationWithUnit.precision(3);
+        averageDurationWithUnit << averageDuration.first << "s";
 
-        if (benchmark.numberOfSimulationRuns > 1)
+        std::ostringstream averageThroughputWithUnit;
+        averageThroughputWithUnit.precision(3);
+        averageThroughputWithUnit << averageThroughput.first << " MiB/s";
+
+        std::ostringstream averageMsgRateWithUnit;
+        averageMsgRateWithUnit << static_cast<int>(averageMsgRate.first) << " 1/s";
+
+        std::cout << std::setw(38) << "- Realtime duration (runtime): " << std::setw(13)
+                  << averageDurationWithUnit.str() << " +/- " << averageDuration.second << "s" << std::endl
+
+                  << std::setw(38) << "- Speedup (virtual time/runtime): " << std::setw(13) << averageSpeedup.first
+                  << " +/- " << averageSpeedup.second << std::endl
+
+                  << std::setw(38) << "- Throughput (data size/runtime): " << std::setw(13)
+                  << averageThroughputWithUnit.str() << " +/- " << averageThroughput.second << " MiB/s" << std::endl
+
+                  << std::setw(38) << "- Message rate (count/runtime): " << std::setw(13)
+                  << averageMsgRateWithUnit.str() << " +/- " << static_cast<int>(averageMsgRate.second) << " 1/s"
+                  << std::endl
+
+                  << std::left << std::setw(38) << "- Total number of messages: " << averageNumberMessages << std::endl
+
+                  << std::endl
+                  << std::endl;
+
+        if (benchmark.writeCsv != "")
         {
-            std::cout << std::left << std::setw(38) << "- Minimum duration: " << *minMaxDuration.first << std::endl;
-            std::cout << std::left << std::setw(38) << "- Maximum duration: " << *minMaxDuration.second << std::endl;
+            std::stringstream csvHeader;
+            csvHeader << "# SilKitBenchmarkDemo, SIL Kit Version " << SilKit::Version::String();
+            const auto csvColumns = "numRuns; participants; messageSize; messageCount; duration(virtual time, s); "
+                                    "numberMessageSent; runtime(s); runtime_err; throughput(MiB/s); "
+                                    "throughput_err; speedup; speedup_err; messageRate(1/s); messageRate_err";
+            std::fstream csvFile;
+            csvFile.open(benchmark.writeCsv, std::ios_base::in | std::ios_base::out); // Try to open
+            bool csvValid{true};
+            if (!csvFile.is_open()) 
+            {
+                // File doesn't exist, create new file and write header
+                csvFile.open(benchmark.writeCsv, std::ios_base::in | std::ios_base::out | std::ios_base::trunc);
+                csvFile << csvHeader.str() << std::endl;
+                csvFile << csvColumns << std::endl;
+            }
+            else
+            {
+                // File is there, check if header is valid
+                std::string header;
+                std::getline(csvFile, header);
+                csvFile.clear();
+                if (header != csvHeader.str())
+                {
+                    std::cerr << "Invalid header in file \"" << benchmark.writeCsv << "\"." << std::endl;
+                    csvValid = false;
+                }
+            }
+            if (csvValid)
+            {
+                // Append data
+                csvFile.seekp(0, std::ios_base::end);
+                csvFile << benchmark.numberOfSimulationRuns << ";" << benchmark.numberOfParticipants << ";"
+                        << benchmark.messageSizeInBytes << ";" << benchmark.messageCount << ";"
+                        << benchmark.simulationDuration.count() << ";" << averageNumberMessages << ";"
+                        << averageDuration.first << ";" << averageDuration.second << ";" << averageThroughput.first
+                        << ";" << averageThroughput.second << ";" << averageSpeedup.first << ";"
+                        << averageSpeedup.second << ";" << averageMsgRate.first << ";" << averageMsgRate.second
+                        << std::endl;
+            }
+            csvFile.close();
+
         }
-        
-        std::cout << std::endl;
+
     }
     catch (const SilKit::ConfigurationError& error)
     {